requirements.txt
xclip
pyperclip
bs4
msal
atexit
requests
python-dateutil
